requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=1.5.3
duckdb>=0.9.0
openai>=1.3.0
plotly>=5.15.0

# File handling
openpyxl>=3.1.2  # For Excel file support
xlrd>=2.0.1      # For older Excel formats

# Utilities
python-dotenv>=1.0.0  # For environment variable management
pytest>=7.4.0    # For testing

# Optional but recommended
numpy>=1.25.2
scikit-learn>=1.3.0  # For basic statistical functions 